Seprating the required functions
